Split scaffolding into parts
