+++ opencog/opencl/scaffolding/show-ocl-hw.cc
@@ -1,0 +1,15 @@
+/**
+ * show-ocl-hw.cc
+ *
+ * Print out OpenCL hardware on a system.
+ *
+ * Copyright (c) 2025 Linas Vepstas
+ */
+
+void report_hardware(void);
+
+/// Print rudimentary report of available OpenCL hardware.
+int main(int argc, char* argv[])
+{
+	report_hardware();
+}
